% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\docType{data}
\name{state_table}
\alias{state_table}
\title{Comex Stat's state table}
\format{
A tibble with 27 rows and 4 columns:
\describe{
\item{state_code}{State numeric code.}
\item{state}{State abbreviation.}
\item{state_name}{State full name in Portuguese.}
\item{region_name}{Brazilian macro region name.}
}
}
\source{
\url{https://balanca.economia.gov.br/balanca/bd/tabelas/UF.csv}
}
\description{
A table containing state codes, abbreviations, names and regions
}
\keyword{internal}
