% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\docType{data}
\name{unit_table}
\alias{unit_table}
\title{Comex Stat's unit code and name table}
\format{
A tibble with 15 rows and 4 columns:
\describe{
\item{unit_code}{NCM unit code.}
\item{unit_name_pt}{Unit name in Portuguese.}
\item{sg_unid}{Unit abbreviation in Portuguese.}
\item{unit_name}{Unit name in English.}
}
}
\source{
\url{https://balanca.economia.gov.br/balanca/bd/tabelas/NCM_UNIDADE.csv}
}
\description{
A table containing unit codes and names
}
\keyword{internal}
